Fix ReScanForeignScan API to make the parameterized query work correctly.
Sub-select or correlated queries that use a parameterized plan
uses ReScanForeignScan API.  However, this API is not correctly
rescanning the data.  In the case of parameters, we need to bind
the new param values after processing them and need to re-execute
the query.  The patch does the same.

In passing, refactor some code in this area and rename a few
functions and variables to match the context.

Reported on GitHub through issues #172 by JackWolfskind
and #201 by lukesilvia.

FDW-157, Suraj Kharage, reviewed by Jeevan Ladhe.

Conflicts:
	expected/select.out
	mysql_fdw.c
	sql/select.sql