Fix crash when sub-select returns NULL.
Sub-select or correlated queries use a parameterized plan, and if the
parameter value is happened to be null because sub-select returned
null; we observed a crash in MySQL lib.  This was because we were
setting the MYSQL_BIND's buffer_type to some valid type and later
also marking the is_null flag to true.  Ideally, this should not
cause any issue, but MySQL lib, seems like, looking at buffer_type
set, tries to read the buffer; ending up with the crash.  To fix
that, set the buffer_type only when we are done with the null
checking i.e. set only when is_null is false.  The issue is exposed
by 7b81a1878e1ed4f3bb06d3e1115f2bdd44d46ad1 that made the
parameterized query work correctly.

Suraj Kharage, reviewed by me, tested by Rajkumar Raghuwanshi.